Daily

+++ ToDoList/ToDoList/Views/ListView.swift
@@ -13,19 +13,29 @@ struct ListView: View {
 
 
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.spring) {
-                            listViewModel.updateItem(item: item)
-                        }
+        ZStack {
+            
+            if listViewModel.items.isEmpty {
+                NoItemsView()
+                    .transition(AnyTransition.opacity.animation(.easeIn))
+            } else {
+                List {
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.spring) {
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
                     }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
+                .listStyle(.plain)
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
+ 
         }
-        .listStyle(.plain)
+  
         .navigationTitle("ToDo List 📝")
         .toolbar {
             ToolbarItem(placement: .automatic) {
